Introduce Invalid CellExpr.
When the (column) expression shifts by a non-zero amount, some cell
expression gets translated into an invalid cell location. Thus, we need
a way to express an invalid cell expression.

However, we haven't worked out on propagating the error to the top level
cell. We should do that in the next commit.