Support adding HTML attributes to the cell.
Note that this is still buggy. Read the comment below.

To make it easier to visualize a larger table, we support adding an HTML
attributes to the cell. However, there are still many bugs:

1. When an ExcelFrame creates a new frame based on it, it doesn't pass
such attribute to the newly created cell. We should first fix the shallow
copy issue then properly resolve this issue.

2. I'd like to use this attribute to the excel as well. To do this,
I'll probably have to have a standalone attribute that can be translated
into both HTML and excel.

3. Adding attributes to the columns isn't supported yet. Maybe they
should be considered as a standalone cell?